Big Update
The goal of this template is to provide players with a framework to easily create reanimation modules through existing animation tools. The update has the main features:
- Player movement enhancements, such as leaning movement, flight, head and torso following mouse, dodging, and more
- A comprehensive list of dances from different games (see FE Dance Player on my profile)
- Animation controller objects that provide the ability to layer animations on top of each other (see Staff Wielder module)
- A module system to allow players to implement and interface their animations (player or accessory) with the controllers of this framework.
- In addition to the module system, players can switch between different animation modules without having to execute a different script.
- R15 Support. The dance list and the module system differentiate between your rig type
This commit also includes an attempted implementation at R6 IK, using the CCDIK Controller. This, however, manipulates existing and additional Motor6ds, the latter of which gets deleted for the main character. I'll leave a CFrame-based implementation to the reader.